% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpconf.R
\name{hpconf}
\alias{hpconf}
\title{Confidence Bands for the Hodrick-Prescott Filter}
\source{
Giles, D. E. (2012) Constructing Confidence Bands for the Hodrick-Prescott Filter. Econometrics Working Papers 1202, University of Victoria.
}
\usage{
hpconf(data, V_y = NULL, ci = 0.05)
}
\arguments{
\item{data}{an object of class "mFilter" containing the output of the function \code{hpfilter}.}

\item{V_y}{a numeric specifying the variance of the cyclical component. If \code{V_y = NULL} (default),
the sample variance of the cyclical component will be used.}

\item{ci}{numeric between 0 and 1 specifying the confidence interval, defaults to 0.05.}
}
\description{
Constructs confidence bands for the Hodrick-Prescott filter
}
