% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hpconf.R
\name{plot.hpconf}
\alias{plot.hpconf}
\title{Plot HP-Filter Confidence Bands}
\usage{
\method{plot}{hpconf}(x, ...)
}
\arguments{
\item{x}{an object of class "hpconf", usually, a result of a call to \code{\link{hpconf}}.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class "hpconf".
}
